Login logic
Changed the launch activity to the Bank acitvity in order to first load the database into the app. Afterwards I call the login activity and request the user information. Created a method in CustomerDatabase that checks if the data inputed returns a user. Created a method in bank activity that provides the customer database with the data it needs to look for. Created a method in login activity to get the responce from the bank weather the user is verified or not.